Move header and add HomeLink component
